chore: renamed back queue -> nats
