+++ sub/internal/transport/nats/publisher/mailer/adapter.go
