+++ sub/internal/transport/nats/publisher/mailer/adapter_test.go
