+++ sub/internal/transport/nats/publisher/mailer/client.go
