+++ sub/internal/transport/nats/publisher/mailer/mocks/mock_publisher.go
